pyproject.toml
[tool.poetry]
name = "electronic-sales"
version = "0.1.0"
description = ""
authors = ["Victor Volkov <vektorn1212@gmail.com>"]
readme = "README.md"
packages = [{include = "electronic_sales"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.1"
django-cors-headers = "^4.3.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
